+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * Correctly assign the type of the actor in a `Command`.
 * Fix error when `TopicListener.stop()` is called and there is not an active connection.
 
+### ✨ Improved
+
+* `Device.start()` now returns `self`.
+
+
 ## 1.1.0 - May 29, 2021
 
 ### 🚀 New